Relax "T" in apodize! and convolve!
Relax "T" in apodize! and convolve! so that eltype of these two arguments can differ.
The original requirement that both have the same type "T" causes issues when trying to use ComplexF32 to inputs to save memory.
It does not seem essential that the two arguments have the same eltype.